==> JSTAR/6_Physical_measures/1_weight.Rmd <==
---
title: "Weight"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_iago.r")
```
<!-- ########################################################## --> 



# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `weight`
* Variable label: `weight of the participant in kg`
* Variable description: `Weight should be collected in kilograms. It should be determined wheather the  variable was collected through a self reported measure or an anthropometric test`
* Domain: `Physical measures`
* Value type: `continuous`
* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

# Data process


### JSTAR - aksst - Wave 1 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `d_005_b`
* Label: `WEIGHT`
* Type of variable: `countinous`

```{r assign_aksst1}
opal.assign.table.tibble(o, 'JSTAR_aksst_w1_weig','JSTAR_datasets_restricted.jstar_2007_5cities',
                         variables=list('d_005_b'), missings = TRUE)
```

```{r local_aksst1}
JSTAR_aksst_w1_weig <- opal.execute(o,'JSTAR_aksst_w1_weig')
Continuous_summary(var = JSTAR_aksst_w1_weig$d_005_b, missing_values = NA)
ggplot(JSTAR_aksst_w1_weig, aes(d_005_b)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`

R code of the ALGORITHM:

```{r harmo_aksst1}
JSTAR_aksst_ds_w1 <- tibble(id=JSTAR_aksst_w1_weig$id)
JSTAR_aksst_ds_w1$weight <- car::recode(JSTAR_aksst_w1_weig$d_005_b, "NA=999")
JSTAR_aksst_ds_w1$weight <- labelled(JSTAR_aksst_ds_w1$weight, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_aksst1}
Continuous_summary(var = JSTAR_aksst_ds_w1$weight, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_aksst_ds_w1, aes(weight)) + geom_histogram(stat="count", fill="steelblue") + xlab("weight") + ylab("Frequency") + xlim(31,168)
```

#### Validation
```{r crosstabulation_aksst1}
```

#### Quality estimation
[Comments on the quality of the new harmonised variable.]

#### Import to OPAL
```{r importOPAL_aksst1}
#val_labels(JSTAR_aksst_ds_w1) <- NULL
#opal.assign.data(o, 'JSTAR_aksst_ds_w1', JSTAR_aksst_ds_w1)
#opal.symbol_import(o,'JSTAR_aksst_ds_w1', project='_Athlos_Harm_Dataset')
weight <- JSTAR_aksst_ds_w1
save(weight, file = "../RData/aksst_w1/weight.RData")
rm(weight)
```







### JSTAR - aksst - Wave 2 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `d_005_b`
* Label: `WEIGHT`
* Type of variable: `countinous`

```{r assign_aksst2}
opal.assign.table.tibble(o, 'JSTAR_aksst_w2_weig','JSTAR_datasets_restricted.jstar_2009_5cities',
                         variables=list('d_005_b'), missings = TRUE)
```

```{r local_aksst2}
JSTAR_aksst_w2_weig <- opal.execute(o,'JSTAR_aksst_w2_weig')
Continuous_summary(var = JSTAR_aksst_w2_weig$d_005_b, missing_values = NA)
ggplot(JSTAR_aksst_w2_weig, aes(d_005_b)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`

R code of the ALGORITHM:

```{r harmo_aksst2}
JSTAR_aksst_ds_w2 <- tibble(id=JSTAR_aksst_w2_weig$id)
JSTAR_aksst_ds_w2$weight <- car::recode(JSTAR_aksst_w2_weig$d_005_b, "NA=999")
JSTAR_aksst_ds_w2$weight <- labelled(JSTAR_aksst_ds_w2$weight, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_aksst2}
Continuous_summary(var = JSTAR_aksst_ds_w2$weight, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_aksst_ds_w2, aes(weight)) + geom_histogram(stat="count", fill="steelblue") + xlab("weight") + ylab("Frequency") + xlim(30,105)
```

#### Validation
```{r crosstabulation_aksst2}
```

#### Quality estimation
[Comments on the quality of the new harmonised variable.]

#### Import to OPAL
```{r importOPAL_aksst2}
#val_labels(JSTAR_aksst_ds_w2) <- NULL
#opal.assign.data(o, 'JSTAR_aksst_ds_w2', JSTAR_aksst_ds_w2)
#opal.symbol_import(o,'JSTAR_aksst_ds_w2', project='_Athlos_Harm_Dataset')
weight <- JSTAR_aksst_ds_w2
save(weight, file = "../RData/aksst_w2/weight.RData")
rm(weight)
```





### JSTAR - aksst - Wave 3 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `d_005_b`
* Label: `WEIGHT`
* Type of variable: `countinous`

```{r assign_aksst3}
opal.assign.table.tibble(o, 'JSTAR_aksst_w3_weig','JSTAR_datasets_restricted.jstar_2011_5cities',
                         variables=list('d_005_b'), missings = TRUE)
```

```{r local_aksst3}
JSTAR_aksst_w3_weig <- opal.execute(o,'JSTAR_aksst_w3_weig')
Continuous_summary(var = JSTAR_aksst_w3_weig$d_005_b, missing_values = NA)
ggplot(JSTAR_aksst_w3_weig, aes(d_005_b)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`

R code of the ALGORITHM:

```{r harmo_aksst3}
JSTAR_aksst_ds_w3 <- tibble(id=JSTAR_aksst_w3_weig$id)
JSTAR_aksst_ds_w3$weight <- car::recode(JSTAR_aksst_w3_weig$d_005_b, "NA=999")
JSTAR_aksst_ds_w3$weight <- labelled(JSTAR_aksst_ds_w3$weight, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_aksst3}
Continuous_summary(var = JSTAR_aksst_ds_w3$weight, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_aksst_ds_w3, aes(weight)) + geom_histogram(stat="count", fill="steelblue") + xlab("weight") + ylab("Frequency") + xlim(31,105)
```

#### Validation
```{r crosstabulation_aksst3}
```

#### Quality estimation
[Comments on the quality of the new harmonised variable.]

#### Import to OPAL
```{r importOPAL_aksst3}
#val_labels(JSTAR_aksst_ds_w3) <- NULL
#opal.assign.data(o, 'JSTAR_aksst_ds_w3', JSTAR_aksst_ds_w3)
#opal.symbol_import(o,'JSTAR_aksst_ds_w3', project='_Athlos_Harm_Dataset')
weight <- JSTAR_aksst_ds_w3
save(weight, file = "../RData/aksst_w3/weight.RData")
rm(weight)
```







### JSTAR - tn - Wave 1 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `d_005_b`
* Label: `WEIGHT`
* Type of variable: `countinous`

```{r assign_tn1}
opal.assign.table.tibble(o, 'JSTAR_tn_w1_weig','JSTAR_datasets_restricted.jstar_2009_2cities',
                         variables=list('d_005_b'), missings = TRUE)
```

```{r local_tn1}
JSTAR_tn_w1_weig <- opal.execute(o,'JSTAR_tn_w1_weig')
Continuous_summary(var = JSTAR_tn_w1_weig$d_005_b, missing_values = NA)
ggplot(JSTAR_tn_w1_weig, aes(d_005_b)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`

R code of the ALGORITHM:

```{r harmo_tn1}
JSTAR_tn_ds_w1 <- tibble(id=JSTAR_tn_w1_weig$id)
JSTAR_tn_ds_w1$weight <- car::recode(JSTAR_tn_w1_weig$d_005_b, "NA=999")
JSTAR_tn_ds_w1$weight <- labelled(JSTAR_tn_ds_w1$weight, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_tn1}
Continuous_summary(var = JSTAR_tn_ds_w1$weight, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_tn_ds_w1, aes(weight)) + geom_histogram(stat="count", fill="steelblue") + xlab("weight") + ylab("Frequency") + xlim(31,130)
```

#### Validation
```{r crosstabulation_tn1}
```

#### Quality estimation
[Comments on the quality of the new harmonised variable.]

#### Import to OPAL
```{r importOPAL_tn1}
#val_labels(JSTAR_tn_ds_w1) <- NULL
#opal.assign.data(o, 'JSTAR_tn_ds_w1', JSTAR_tn_ds_w1)
#opal.symbol_import(o,'JSTAR_tn_ds_w1', project='_Athlos_Harm_Dataset')
weight <- JSTAR_tn_ds_w1
save(weight, file = "../RData/tn_w1/weight.RData")
rm(weight)
```





### JSTAR - tn - Wave 2 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `d_005_b`
* Label: `WEIGHT`
* Type of variable: `countinous`

```{r assign_tn2}
opal.assign.table.tibble(o, 'JSTAR_tn_w2_weig','JSTAR_datasets_restricted.jstar_2011_2cities',
                         variables=list('d_005_b'), missings = TRUE)
```

```{r local_tn2}
JSTAR_tn_w2_weig <- opal.execute(o,'JSTAR_tn_w2_weig')
Continuous_summary(var = JSTAR_tn_w2_weig$d_005_b, missing_values = NA)
ggplot(JSTAR_tn_w2_weig, aes(d_005_b)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`

R code of the ALGORITHM:

```{r harmo_tn2}
JSTAR_tn_ds_w2 <- tibble(id=JSTAR_tn_w2_weig$id)
JSTAR_tn_ds_w2$weight <- car::recode(JSTAR_tn_w2_weig$d_005_b, "NA=999")
JSTAR_tn_ds_w2$weight <- labelled(JSTAR_tn_ds_w2$weight, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_tn2}
Continuous_summary(var = JSTAR_tn_ds_w2$weight, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_tn_ds_w2, aes(weight)) + geom_histogram(stat="count", fill="steelblue") + xlab("weight") + ylab("Frequency") + xlim(30,154)
```

#### Validation
```{r crosstabulation_tn2}
```

#### Quality estimation
[Comments on the quality of the new harmonised variable.]

#### Import to OPAL
```{r importOPAL_tn2}
#val_labels(JSTAR_tn_ds_w2) <- NULL
#opal.assign.data(o, 'JSTAR_tn_ds_w2', JSTAR_tn_ds_w2)
#opal.symbol_import(o,'JSTAR_tn_ds_w2', project='_Athlos_Harm_Dataset')
weight <- JSTAR_tn_ds_w2
save(weight, file = "../RData/tn_w2/weight.RData")
rm(weight)
```





### JSTAR - cth - Wave 1 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `d_005_b`
* Label: `WEIGHT`
* Type of variable: `countinous`

```{r assign_cth1}
opal.assign.table.tibble(o, 'JSTAR_cth_w1_weig','JSTAR_datasets_restricted.jstar_2011_3cities',
                         variables=list('d_005_b'), missings = TRUE)
```

```{r local_cth1}
JSTAR_cth_w1_weig <- opal.execute(o,'JSTAR_cth_w1_weig')
Continuous_summary(var = JSTAR_cth_w1_weig$d_005_b, missing_values = NA)
ggplot(JSTAR_cth_w1_weig, aes(d_005_b)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`

R code of the ALGORITHM:

```{r harmo_cth1}
JSTAR_cth_ds_w1 <- tibble(id=JSTAR_cth_w1_weig$id)
JSTAR_cth_ds_w1$weight <- car::recode(JSTAR_cth_w1_weig$d_005_b, "NA=999")
JSTAR_cth_ds_w1$weight <- labelled(JSTAR_cth_ds_w1$weight, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_cth1}
Continuous_summary(var = JSTAR_cth_ds_w1$weight, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_cth_ds_w1, aes(weight)) + geom_histogram(stat="count", fill="steelblue") + xlab("weight") + ylab("Frequency") + xlim(25,100)
```

#### Validation
There is an unusual value that is lower than 30
```{r crosstabulation_cth1}
JSTAR_cth_w1_weig[which(JSTAR_cth_w1_weig$d_005_b<30),]
```

#### Quality estimation
No comments.



#### Import to OPAL
```{r importOPAL_cth1}
#val_labels(JSTAR_cth_ds_w1) <- NULL
#opal.assign.data(o, 'JSTAR_cth_ds_w1', JSTAR_cth_ds_w1)
#opal.symbol_import(o,'JSTAR_cth_ds_w1', project='_Athlos_Harm_Dataset')
weight <- JSTAR_cth_ds_w1
save(weight, file = "../RData/cth_w1/weight.RData")
rm(weight)
```

```{r closeRsession, echo=FALSE}
opal.logout(o)
```
