==> JSTAR/6_Physical_measures/3_BMI.Rmd <==
---
title: "BMI"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_iago.r")
```
<!-- ########################################################## --> 



# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `bmi`
* Variable label: `Body Mass Index calculated as weight (kg)/height(m)^2`
* Variable description: `the BMI variable should be created by dividing weight and height variables with the following equation= weight (kg)/height(m)^2`
* Domain: `Physical measures`
* Value type: `continuous`
* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

# Data process


```{r lab, echo=FALSE}
ds_label <- "Body Mass Index calculated as weight (kg)/height(m)^2"
```

### JSTAR - aksst - Wave 1 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `d_005_b`
* Label: `WEIGHT`
* Type of variable: `countinous`

* Name: `d_006_b`
* Label: `HEIGHT`
* Type of variable: `countinous`

```{r assign_aksst1}
opal.assign.table.tibble(o, 'JSTAR_aksst_w1_bmi','JSTAR_datasets_restricted.jstar_2007_5cities',
                         variables=list('d_005_b', 'd_006_b'), missings = TRUE)
```

```{r local_aksst1}
JSTAR_aksst_w1_bmi <- opal.execute(o,'JSTAR_aksst_w1_bmi')

Continuous_summary(var = JSTAR_aksst_w1_bmi$d_005_b, missing_values = NA)
ggplot(JSTAR_aksst_w1_bmi, aes(d_005_b)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency")

Continuous_summary(var = JSTAR_aksst_w1_bmi$d_006_b, missing_values = c(888,999))
ggplot(JSTAR_aksst_w1_bmi, aes(d_006_b)) + geom_histogram(stat="count", fill="steelblue") + xlab("Height") + ylab("Frequency") 
```

#### Harmonisation algorithm
The harmonized variable will be calculate from the study-specific variables using this formula

bmi=weight/height^2

The missing values have to be recoded as follows:

* `NA into 999`

R code of the ALGORITHM:

```{r harmo_aksst1}
JSTAR_aksst_ds_w1 <- tibble(id=JSTAR_aksst_w1_bmi$id)
JSTAR_aksst_ds_w1$bmi <- car::recode((JSTAR_aksst_w1_bmi$d_005_b / (JSTAR_aksst_w1_bmi$d_006_b/100)^2), "NA=999")
JSTAR_aksst_ds_w1$bmi <- labelled(JSTAR_aksst_ds_w1$bmi, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_aksst1}
Continuous_summary(var = JSTAR_aksst_ds_w1$bmi, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_aksst_ds_w1, aes(bmi)) + geom_histogram(stat="count", fill="steelblue") + xlab("BMI") + ylab("Frequency") + xlim(10,100)
```

#### Validation
```{r crosstabulation_aksst1}
```

#### Quality estimation
[Comments on the quality of the new harmonised variable.]

#### Import to OPAL
```{r importOPAL_aksst1}
#opal.assign.data(o, 'JSTAR_aksst_ds_w1', JSTAR_aksst_ds_w1)
#opal.symbol_import(o,'JSTAR_aksst_ds_w1', project='_Athlos_Harm_Dataset')
bmi <- JSTAR_aksst_ds_w1
save(bmi, file = "../RData/aksst_w1/bmi.RData")
rm(bmi)

JSTAR_aksst_m_ds_w1 <- JSTAR_aksst_ds_w1
JSTAR_aksst_m_ds_w1$bmi <- car::recode(JSTAR_aksst_m_ds_w1$bmi, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'JSTAR_aksst_m_ds_w1', JSTAR_aksst_m_ds_w1)
#opal.symbol_import(o,'JSTAR_aksst_m_ds_w1', project='_Athlos_Harm_Dataset')
```







### JSTAR - aksst - Wave 2 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `d_005_b`
* Label: `WEIGHT`
* Type of variable: `countinous`

* Name: `d_006_b`
* Label: `HEIGHT`
* Type of variable: `countinous`


```{r assign_aksst2}
opal.assign.table.tibble(o, 'JSTAR_aksst_w2_bmi','JSTAR_datasets_restricted.jstar_2009_5cities',
                         variables=list('d_005_b', 'd_006_b'), missings = TRUE)
```

```{r local_aksst2}
JSTAR_aksst_w2_bmi <- opal.execute(o,'JSTAR_aksst_w2_bmi')

Continuous_summary(var = JSTAR_aksst_w2_bmi$d_005_b, missing_values = NA)
ggplot(JSTAR_aksst_w2_bmi, aes(d_005_b)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency")

Continuous_summary(var = JSTAR_aksst_w2_bmi$d_006_b, missing_values = NA)
ggplot(JSTAR_aksst_w2_bmi, aes(d_006_b)) + geom_histogram(stat="count", fill="steelblue") + xlab("Height") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable will be calculate from the study-specific variables using this formula

bmi=weight/height^2

The missing values have to be recoded as follows:

* `NA into 999`

R code of the ALGORITHM:

```{r harmo_aksst2}
JSTAR_aksst_ds_w2 <- tibble(id=JSTAR_aksst_w2_bmi$id)
JSTAR_aksst_ds_w2$bmi <- car::recode((JSTAR_aksst_w2_bmi$d_005_b / (JSTAR_aksst_w2_bmi$d_006_b/100)^2), "NA=999")
JSTAR_aksst_ds_w2$bmi <- labelled(JSTAR_aksst_ds_w2$bmi, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_aksst2}
Continuous_summary(var = JSTAR_aksst_ds_w2$bmi, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_aksst_ds_w2, aes(bmi)) + geom_histogram(stat="count", fill="steelblue") + xlab("BMI") + ylab("Frequency") + xlim(10,100)
```

#### Validation
```{r crosstabulation_aksst2}
```

#### Quality estimation
[Comments on the quality of the new harmonised variable.]

#### Import to OPAL
```{r importOPAL_aksst2}
#opal.assign.data(o, 'JSTAR_aksst_ds_w2', JSTAR_aksst_ds_w2)
#opal.symbol_import(o,'JSTAR_aksst_ds_w2', project='_Athlos_Harm_Dataset')
bmi <- JSTAR_aksst_ds_w2
save(bmi, file = "../RData/aksst_w2/bmi.RData")
rm(bmi)

JSTAR_aksst_m_ds_w2 <- JSTAR_aksst_ds_w2
JSTAR_aksst_m_ds_w2$bmi <- car::recode(JSTAR_aksst_m_ds_w2$bmi, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'JSTAR_aksst_m_ds_w2', JSTAR_aksst_m_ds_w2)
#opal.symbol_import(o,'JSTAR_aksst_m_ds_w2', project='_Athlos_Harm_Dataset')
```





### JSTAR - aksst - Wave 3 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `d_005_b`
* Label: `WEIGHT`
* Type of variable: `countinous`

* Name: `d_006_b`
* Label: `HEIGHT`
* Type of variable: `countinous`


```{r assign_aksst3}
opal.assign.table.tibble(o, 'JSTAR_aksst_w3_bmi','JSTAR_datasets_restricted.jstar_2011_5cities',
                         variables=list('d_005_b', 'd_006_b'), missings = TRUE)
```

```{r local_aksst3}
JSTAR_aksst_w3_bmi <- opal.execute(o,'JSTAR_aksst_w3_bmi')

Continuous_summary(var = JSTAR_aksst_w3_bmi$d_005_b, missing_values = NA)
ggplot(JSTAR_aksst_w3_bmi, aes(d_005_b)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency")

Continuous_summary(var = JSTAR_aksst_w3_bmi$d_006_b, missing_values = c(888,999))
ggplot(JSTAR_aksst_w3_bmi, aes(d_006_b)) + geom_histogram(stat="count", fill="steelblue") + xlab("Height") + ylab("Frequency") 
```

#### Harmonisation algorithm
The harmonized variable will be calculate from the study-specific variables using this formula

bmi=weight/height^2

The missing values have to be recoded as follows:

* `NA into 999`

R code of the ALGORITHM:

```{r harmo_aksst3}
JSTAR_aksst_ds_w3 <- tibble(id=JSTAR_aksst_w3_bmi$id)
JSTAR_aksst_ds_w3$bmi <- car::recode((JSTAR_aksst_w3_bmi$d_005_b / (JSTAR_aksst_w3_bmi$d_006_b/100)^2), "NA=999")
JSTAR_aksst_ds_w3$bmi <- labelled(JSTAR_aksst_ds_w3$bmi, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_aksst3}
Continuous_summary(var = JSTAR_aksst_ds_w3$bmi, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_aksst_ds_w3, aes(bmi)) + geom_histogram(stat="count", fill="steelblue") + xlab("BMI") + ylab("Frequency") + xlim(10,100)
```

#### Validation
```{r crosstabulation_aksst3}
```

#### Quality estimation
[Comments on the quality of the new harmonised variable.]

#### Import to OPAL
```{r importOPAL_aksst3}
#opal.assign.data(o, 'JSTAR_aksst_ds_w3', JSTAR_aksst_ds_w3)
#opal.symbol_import(o,'JSTAR_aksst_ds_w3', project='_Athlos_Harm_Dataset')
bmi <- JSTAR_aksst_ds_w3
save(bmi, file = "../RData/aksst_w3/bmi.RData")
rm(bmi)

JSTAR_aksst_m_ds_w3 <- JSTAR_aksst_ds_w3
JSTAR_aksst_m_ds_w3$bmi <- car::recode(JSTAR_aksst_m_ds_w3$bmi, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'JSTAR_aksst_m_ds_w3', JSTAR_aksst_m_ds_w3)
#opal.symbol_import(o,'JSTAR_aksst_m_ds_w3', project='_Athlos_Harm_Dataset')
```






### JSTAR - tn - Wave 1 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `d_005_b`
* Label: `WEIGHT`
* Type of variable: `countinous`

* Name: `d_006_b`
* Label: `HEIGHT`
* Type of variable: `countinous`


```{r assign_tn1}
opal.assign.table.tibble(o, 'JSTAR_tn_w1_bmi','JSTAR_datasets_restricted.jstar_2009_2cities',
                         variables=list('d_005_b', 'd_006_b'), missings = TRUE)
```

```{r local_tn1}
JSTAR_tn_w1_bmi <- opal.execute(o,'JSTAR_tn_w1_bmi')

Continuous_summary(var = JSTAR_tn_w1_bmi$d_005_b, missing_values = NA)
ggplot(JSTAR_tn_w1_bmi, aes(d_005_b)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency")

Continuous_summary(var = JSTAR_tn_w1_bmi$d_006_b, missing_values = c(888,999))
ggplot(JSTAR_tn_w1_bmi, aes(d_006_b)) + geom_histogram(stat="count", fill="steelblue") + xlab("Height") + ylab("Frequency") 
```

#### Harmonisation algorithm
The harmonized variable will be calculate from the study-specific variables using this formula

bmi=weight/height^2

The missing values have to be recoded as follows:

* `NA into 999`

R code of the ALGORITHM:

```{r harmo_tn1}
JSTAR_tn_ds_w1 <- tibble(id=JSTAR_tn_w1_bmi$id)
JSTAR_tn_ds_w1$bmi <- car::recode((JSTAR_tn_w1_bmi$d_005_b / (JSTAR_tn_w1_bmi$d_006_b/100)^2), "NA=999")
JSTAR_tn_ds_w1$bmi <- labelled(JSTAR_tn_ds_w1$bmi, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_tn1}
Continuous_summary(var = JSTAR_tn_ds_w1$bmi, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_tn_ds_w1, aes(bmi)) + geom_histogram(stat="count", fill="steelblue") + xlab("BMI") + ylab("Frequency") + xlim(10,100)
```

#### Validation
```{r crosstabulation_tn1}
```

#### Quality estimation
[Comments on the quality of the new harmonised variable.]

#### Import to OPAL
```{r importOPAL_tn1}
#opal.assign.data(o, 'JSTAR_tn_ds_w1', JSTAR_tn_ds_w1)
#opal.symbol_import(o,'JSTAR_tn_ds_w1', project='_Athlos_Harm_Dataset')
bmi <- JSTAR_tn_ds_w1
save(bmi, file = "../RData/tn_w1/bmi.RData")
rm(bmi)

JSTAR_tn_m_ds_w1 <- JSTAR_tn_ds_w1
JSTAR_tn_m_ds_w1$bmi <- car::recode(JSTAR_tn_m_ds_w1$bmi, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'JSTAR_tn_m_ds_w1', JSTAR_tn_m_ds_w1)
#opal.symbol_import(o,'JSTAR_tn_m_ds_w1', project='_Athlos_Harm_Dataset')
```





### JSTAR - tn - Wave 2 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `d_005_b`
* Label: `WEIGHT`
* Type of variable: `countinous`

* Name: `d_006_b`
* Label: `HEIGHT`
* Type of variable: `countinous`

```{r assign_tn2}
opal.assign.table.tibble(o, 'JSTAR_tn_w2_bmi','JSTAR_datasets_restricted.jstar_2011_2cities',
                         variables=list('d_005_b', 'd_006_b'), missings = TRUE)
```

```{r local_tn2}
JSTAR_tn_w2_bmi <- opal.execute(o,'JSTAR_tn_w2_bmi')

Continuous_summary(var = JSTAR_tn_w2_bmi$d_005_b, missing_values = NA)
ggplot(JSTAR_tn_w2_bmi, aes(d_005_b)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency")

Continuous_summary(var = JSTAR_tn_w2_bmi$d_006_b, missing_values = c(888,999))
ggplot(JSTAR_tn_w2_bmi, aes(d_006_b)) + geom_histogram(stat="count", fill="steelblue") + xlab("Height") + ylab("Frequency") 
```

#### Harmonisation algorithm
The harmonized variable will be calculate from the study-specific variables using this formula

bmi=weight/height^2

The missing values have to be recoded as follows:

* `NA into 999`

R code of the ALGORITHM:

```{r harmo_tn2}
JSTAR_tn_ds_w2 <- tibble(id=JSTAR_tn_w2_bmi$id)
JSTAR_tn_ds_w2$bmi <- car::recode((JSTAR_tn_w2_bmi$d_005_b / (JSTAR_tn_w2_bmi$d_006_b/100)^2), "NA=999")
JSTAR_tn_ds_w2$bmi <- labelled(JSTAR_tn_ds_w2$bmi, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_tn2}
Continuous_summary(var = JSTAR_tn_ds_w2$bmi, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_tn_ds_w2, aes(bmi)) + geom_histogram(stat="count", fill="steelblue") + xlab("BMI") + ylab("Frequency") + xlim(10,100)
```

#### Validation
```{r crosstabulation_tn2}
```

#### Quality estimation
[Comments on the quality of the new harmonised variable.]

#### Import to OPAL
```{r importOPAL_tn2}
#opal.assign.data(o, 'JSTAR_tn_ds_w2', JSTAR_tn_ds_w2)
#opal.symbol_import(o,'JSTAR_tn_ds_w2', project='_Athlos_Harm_Dataset')
bmi <- JSTAR_tn_ds_w2
save(bmi, file = "../RData/tn_w2/bmi.RData")
rm(bmi)

JSTAR_tn_m_ds_w2 <- JSTAR_tn_ds_w2
JSTAR_tn_m_ds_w2$bmi <- car::recode(JSTAR_tn_m_ds_w2$bmi, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'JSTAR_tn_m_ds_w2', JSTAR_tn_m_ds_w2)
#opal.symbol_import(o,'JSTAR_tn_m_ds_w2', project='_Athlos_Harm_Dataset')
```





### JSTAR - cth - Wave 1 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `d_005_b`
* Label: `WEIGHT`
* Type of variable: `countinous`

* Name: `d_006_b`
* Label: `HEIGHT`
* Type of variable: `countinous`

```{r assign_cth1}
opal.assign.table.tibble(o, 'JSTAR_cth_w1_bmi','JSTAR_datasets_restricted.jstar_2011_3cities',
                         variables=list('d_005_b', 'd_006_b'), missings = TRUE)
```

```{r local_cth1}
JSTAR_cth_w1_bmi <- opal.execute(o,'JSTAR_cth_w1_bmi')

Continuous_summary(var = JSTAR_cth_w1_bmi$d_005_b, missing_values = NA)
ggplot(JSTAR_cth_w1_bmi, aes(d_005_b)) + geom_histogram(stat="count", fill="steelblue") + xlab("Weight") + ylab("Frequency")

Continuous_summary(var = JSTAR_cth_w1_bmi$d_006_b, missing_values = c(888,999))
ggplot(JSTAR_cth_w1_bmi, aes(d_006_b)) + geom_histogram(stat="count", fill="steelblue") + xlab("Height") + ylab("Frequency") 
```

#### Harmonisation algorithm
The harmonized variable will be calculate from the study-specific variables using this formula

bmi=weight/height^2

The missing values have to be recoded as follows:

* `NA into 999`

R code of the ALGORITHM:

```{r harmo_cth1}
JSTAR_cth_ds_w1 <- tibble(id=JSTAR_cth_w1_bmi$id)
JSTAR_cth_ds_w1$bmi <- car::recode((JSTAR_cth_w1_bmi$d_005_b / (JSTAR_cth_w1_bmi$d_006_b/100)^2), "NA=999")
JSTAR_cth_ds_w1$bmi <- labelled(JSTAR_cth_ds_w1$bmi, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_cth1}
Continuous_summary(var = JSTAR_cth_ds_w1$bmi, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_cth_ds_w1, aes(bmi)) + geom_histogram(stat="count", fill="steelblue") + xlab("BMI") + ylab("Frequency") + xlim(7,40)
```



#### Validation
There is an unusual value that is lower than 12:
```{r crosstabulation_cth1}
JSTAR_cth_w1_bmi$bmi <- JSTAR_cth_ds_w1$bmi
JSTAR_cth_w1_bmi[which(JSTAR_cth_w1_bmi$bmi<13),]
```

#### Quality estimation
No comments.

#### Import to OPAL
```{r importOPAL_cth1}
#opal.assign.data(o, 'JSTAR_cth_ds_w1', JSTAR_cth_ds_w1)
#opal.symbol_import(o,'JSTAR_cth_ds_w1', project='_Athlos_Harm_Dataset')
bmi <- JSTAR_cth_ds_w1
save(bmi, file = "../RData/cth_w1/bmi.RData")
rm(bmi)

JSTAR_cth_m_ds_w1 <- JSTAR_cth_ds_w1
JSTAR_cth_m_ds_w1$bmi <- car::recode(JSTAR_cth_m_ds_w1$bmi, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'JSTAR_cth_m_ds_w1', JSTAR_cth_m_ds_w1)
#opal.symbol_import(o,'JSTAR_cth_m_ds_w1', project='_Athlos_Harm_Dataset')
```

```{r closeRsession, echo=FALSE}
opal.logout(o)
```
