==> JSTAR/6_Physical_measures/4_Obesity.Rmd <==
---
title: "Obesity"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ############################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_iago.r")
```
<!-- ############################## --> 



# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `obesity`
* Variable label: `Presence of obesity based in Body Mass Index classification according to the WHO BMI classification (http://apps.who.int/bmi/index.jsp?introPage=intro_3.html): BMI value equal to or above 30Kg/m^2`
* Variable description: `the BMI variable should be created by dividing weight and hiegh variables by means of the following equation= weight (kg)/height(m)^2. Then, the Obesity variable should be generated according to the WHO cutoff point for BMI classification`
* Domain: `Physical measures`
* Value type: `categorical`
* Category coding:

**Code** | **Category Label**
-------- | ------------------
0        | not obese
1        | obese

* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

# Data process

```{r lab, echo=FALSE}
ds_label <- "Precense of obesity based in Body Mass Index classification according to the WHO BMI classification (http://apps.who.int/bmi/index.jsp?introPage=intro_3.html): BMI value equal to or above 30Kg/m^2"
```


## Adachi-Kanazawa-Shirakawa-Sendai-Takikawa

### Wave 1

#### Study-specific variable description

The study-specific variables elected to be harmonised are:

* Name: `bmi`
* Label: `Body Mass Index calculated as weight (kg)/height(m)^2`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not applicable`
    + `995 = Does not answer`
    + `996 = Not attempt/not done`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`
    
```{r assign_aksst1, include=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/JSTAR/RData/aksst_w1/bmi.RData")
JSTAR_bmi <- bmi
```

```{r local_aksst1}
Continuous_summary(JSTAR_bmi[[2]], missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_bmi, aes(x=factor(JSTAR_bmi[[2]]))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("BMI") + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

```{r harmo_aksst1}
JSTAR_aksst_ds_w1 <- tibble(id=JSTAR_bmi$id)
JSTAR_aksst_ds_w1$obesity <- car::recode(JSTAR_bmi[[2]], "lo:29.9999999=0; 30:900=1; NA=999")
JSTAR_aksst_ds_w1$obesity <- labelled(JSTAR_aksst_ds_w1[[2]], labels = c("not obese"=0, "obese"=1,  "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript_aksst1}
Categorical_summary(var = JSTAR_aksst_ds_w1[[2]], missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_aksst_ds_w1, aes(x=factor(JSTAR_aksst_ds_w1[[2]]))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Obesity") + ylab("Frequency")
```

#### Validation
```{r crosstab_aksst1}
BeforeH <- c(sum(na.omit(JSTAR_bmi[[2]])<30),sum(na.omit(JSTAR_bmi[[2]])>=30 & na.omit(JSTAR_bmi[[2]])<=900), sum(na.omit(JSTAR_bmi[[2]])==999))
AfterH <- table(JSTAR_aksst_ds_w1[[2]], useNA = "ifany")
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("<30->0",">=30->1","999") 
kable(C)
```

#### Quality estimation
[Comments on the quality of the new harmonised variable. This section can be done in a later stage, once all variables are harmonised.]

```{r read_aksst1, include=FALSE}
rm(JSTAR_bmi)
```


### Wave 2

#### Study-specific variable description

The study-specific variables elected to be harmonised are:

* Name: `bmi`
* Label: `Body Mass Index calculated as weight (kg)/height(m)^2`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not applicable`
    + `995 = Does not answer`
    + `996 = Not attempt/not done`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`
    
```{r assign_aksst2}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/JSTAR/RData/aksst_w2/bmi.RData")
JSTAR_bmi <- bmi
```

```{r local_aksst2}
Continuous_summary(JSTAR_bmi[[2]], missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_bmi, aes(x=factor(JSTAR_bmi[[2]]))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("BMI") + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

```{r harmo_aksst2}
JSTAR_aksst_ds_w2 <- tibble(id=JSTAR_bmi$id)
JSTAR_aksst_ds_w2$obesity <- car::recode(JSTAR_bmi[[2]], "lo:29.9999999=0; 30:900=1; NA=999")
JSTAR_aksst_ds_w2$obesity <- labelled(JSTAR_aksst_ds_w2[[2]], labels = c("not obese"=0, "obese"=1,  "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript_aksst2}
Categorical_summary(var = JSTAR_aksst_ds_w2[[2]], missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_aksst_ds_w2, aes(x=factor(JSTAR_aksst_ds_w2[[2]]))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Obesity") + ylab("Frequency")
```

#### Validation
```{r crosstab_aksst2}
BeforeH <- c(sum(na.omit(JSTAR_bmi[[2]])<30),sum(na.omit(JSTAR_bmi[[2]])>=30 & na.omit(JSTAR_bmi[[2]])<=900), sum(na.omit(JSTAR_bmi[[2]])==999))
AfterH <- table(JSTAR_aksst_ds_w2[[2]], useNA = "ifany")
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("<30->0",">=30->1","999") 
kable(C)
```

#### Quality estimation
[Comments on the quality of the new harmonised variable. This section can be done in a later stage, once all variables are harmonised.]

```{r read_aksst2, include=FALSE}
rm(JSTAR_bmi)
```


### Wave 3

#### Study-specific variable description

The study-specific variables elected to be harmonised are:

* Name: `bmi`
* Label: `Body Mass Index calculated as weight (kg)/height(m)^2`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not applicable`
    + `995 = Does not answer`
    + `996 = Not attempt/not done`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`

```{r assign_aksst3}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/JSTAR/RData/aksst_w3/bmi.RData")
JSTAR_bmi <- bmi
```

```{r local_aksst3}
Continuous_summary(JSTAR_bmi[[2]], missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_bmi, aes(x=JSTAR_bmi[[2]])) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("BMI") + ylab("Frequency") + xlim(10, 45)
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

```{r harmo_aksst3}
JSTAR_aksst_ds_w3 <- tibble(id=JSTAR_bmi$id)
JSTAR_aksst_ds_w3$obesity <- car::recode(JSTAR_bmi[[2]], "lo:29.9999999=0; 30:900=1; NA=999")
JSTAR_aksst_ds_w3$obesity <- labelled(JSTAR_aksst_ds_w3[[2]], labels = c("not obese"=0, "obese"=1,  "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript_aksst3}
Categorical_summary(var = JSTAR_aksst_ds_w3[[2]], missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_aksst_ds_w3, aes(x=factor(JSTAR_aksst_ds_w3[[2]]))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Obesity") + ylab("Frequency")
```

#### Validation
```{r crosstab_aksst3}
BeforeH <- c(sum(na.omit(JSTAR_bmi[[2]])<30),sum(na.omit(JSTAR_bmi[[2]])>=30 & na.omit(JSTAR_bmi[[2]])<=900), sum(na.omit(JSTAR_bmi[[2]])==999))
AfterH <- table(JSTAR_aksst_ds_w3[[2]], useNA = "ifany")
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("<30->0",">=30->1", "999") 
kable(C)
```

#### Quality estimation
[Comments on the quality of the new harmonised variable. This section can be done in a later stage, once all variables are harmonised.]


## Tosu-Naha

### Wave 1

#### Study-specific variable description

The study-specific variables elected to be harmonised are:

* Name: `bmi`
* Label: `Body Mass Index calculated as weight (kg)/height(m)^2`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not applicable`
    + `995 = Does not answer`
    + `996 = Not attempt/not done`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`
    
```{r assign_tn1, include=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/JSTAR/RData/tn_w1/bmi.RData")
JSTAR_bmi <- bmi
```

```{r local_tn1}
Continuous_summary(JSTAR_bmi[[2]], missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_bmi, aes(x=factor(JSTAR_bmi[[2]]))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("BMI") + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

```{r harmo_tn1}
JSTAR_tn_ds_w1 <- tibble(id=JSTAR_bmi$id)
JSTAR_tn_ds_w1$obesity <- car::recode(JSTAR_bmi[[2]], "lo:29.9999999=0; 30:900=1; NA=999")
JSTAR_tn_ds_w1$obesity <- labelled(JSTAR_tn_ds_w1[[2]], labels = c("not obese"=0, "obese"=1,  "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript_tn1}
Categorical_summary(var = JSTAR_tn_ds_w1[[2]], missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_tn_ds_w1, aes(x=factor(JSTAR_tn_ds_w1[[2]]))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Obesity") + ylab("Frequency")
```

#### Validation
```{r crosstab_tn1}
BeforeH <- c(sum(na.omit(JSTAR_bmi[[2]])<30),sum(na.omit(JSTAR_bmi[[2]])>=30 & na.omit(JSTAR_bmi[[2]])<=900), sum(c(sum(na.omit(JSTAR_bmi[[2]])==999), sum(is.na(JSTAR_bmi[[2]])) ) ))
AfterH <- table(JSTAR_tn_ds_w1[[2]], useNA = "ifany")
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("<30->0",">=30->1","999") 
kable(C)
```

#### Quality estimation
[Comments on the quality of the new harmonised variable. This section can be done in a later stage, once all variables are harmonised.]

```{r read_tn, include=FALSE}
rm(JSTAR_bmi)
```


### Wave 2

#### Study-specific variable description

The study-specific variables elected to be harmonised are:

* Name: `bmi`
* Label: `Body Mass Index calculated as weight (kg)/height(m)^2`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not applicable`
    + `995 = Does not answer`
    + `996 = Not attempt/not done`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`
    
```{r assign_tn2}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/JSTAR/RData/tn_w2/bmi.RData")
JSTAR_bmi <- bmi
```

```{r local_tn2}
Continuous_summary(JSTAR_bmi[[2]], missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_bmi, aes(x=factor(JSTAR_bmi[[2]]))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("BMI") + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

```{r harmo_tn2}
JSTAR_tn_ds_w2 <- tibble(id=JSTAR_bmi$id)
JSTAR_tn_ds_w2$obesity <- car::recode(JSTAR_bmi[[2]], "lo:29.9999999=0; 30:900=1; NA=999")
JSTAR_tn_ds_w2$obesity <- labelled(JSTAR_tn_ds_w2[[2]], labels = c("not obese"=0, "obese"=1,  "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript_tn2}
Categorical_summary(var = JSTAR_tn_ds_w2[[2]], missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_tn_ds_w2, aes(x=factor(JSTAR_tn_ds_w2[[2]]))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Obesity") + ylab("Frequency")
```

#### Validation
```{r crosstab_tn2}
BeforeH <- c(sum(na.omit(JSTAR_bmi[[2]])<30),sum(na.omit(JSTAR_bmi[[2]])>=30 & na.omit(JSTAR_bmi[[2]])<=900), sum(na.omit(JSTAR_bmi[[2]])==999))
AfterH <- table(JSTAR_tn_ds_w2[[2]], useNA = "ifany")
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("<30->0",">=30->1","999") 
kable(C)
```

#### Quality estimation
[Comments on the quality of the new harmonised variable. This section can be done in a later stage, once all variables are harmonised.]

```{r read_tn2, include=FALSE}
rm(JSTAR_bmi)
```


## Chofu-Tondabayashi-Hiroshima

### Wave 1

#### Study-specific variable description

The study-specific variables elected to be harmonised are:

* Name: `bmi`
* Label: `Body Mass Index calculated as weight (kg)/height(m)^2`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not applicable`
    + `995 = Does not answer`
    + `996 = Not attempt/not done`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`
    
```{r assign_cth1, include=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/JSTAR/RData/cth_w1/bmi.RData")
JSTAR_bmi <- bmi
```

```{r local_cth1}
Continuous_summary(JSTAR_bmi[[2]], missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_bmi, aes(x=factor(JSTAR_bmi[[2]]))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("BMI") + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

```{r harmo_cth1}
JSTAR_cth_ds_w1 <- tibble(id=JSTAR_bmi$id)
JSTAR_cth_ds_w1$obesity <- car::recode(JSTAR_bmi[[2]], "lo:29.9999999=0; 30:900=1; NA=999")
JSTAR_cth_ds_w1$obesity <- labelled(JSTAR_cth_ds_w1[[2]], labels = c("not obese"=0, "obese"=1,  "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript_cth1}
Categorical_summary(var = JSTAR_cth_ds_w1[[2]], missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_cth_ds_w1, aes(x=factor(JSTAR_cth_ds_w1[[2]]))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Obesity") + ylab("Frequency")
```

#### Validation
```{r crosstab_cth1}
BeforeH <- c(sum(na.omit(JSTAR_bmi[[2]])<30),sum(na.omit(JSTAR_bmi[[2]])>=30 & na.omit(JSTAR_bmi[[2]])<=900), sum(na.omit(JSTAR_bmi[[2]])==999))
AfterH <- table(JSTAR_cth_ds_w1[[2]], useNA = "ifany")
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("<30->0",">=30->1","999") 
kable(C)
```

#### Quality estimation
Once BMI has been calculated using weight (Kg) and Height (m^2),
Obesity was calculated by using the BMI cutoff>=30kg/m2 following the WHO 
classification. Finally, the resulting BMI variable was coded as follows: 
BMI>30kg/m2 =1 (obese); BMI<30kg/m2 =0 (not obese) to generate two categories.


```{r read_cth1, include=FALSE}
rm(JSTAR_bmi)
```

<!--- ######################### Import to OPAL #######################-->
```{r importOPAL, echo=FALSE}
obesity <- JSTAR_aksst_m_ds_w1 <- JSTAR_aksst_ds_w1
save(obesity, file = "../RData/aksst_w1/obesity.RData")
rm(obesity)

obesity <- JSTAR_aksst_m_ds_w2   <- JSTAR_aksst_ds_w2
save(obesity, file = "../RData/aksst_w2/obesity.RData")
rm(obesity)

obesity <- JSTAR_aksst_m_ds_w3   <- JSTAR_aksst_ds_w3
save(obesity, file = "../RData/aksst_w3/obesity.RData")
rm(obesity)


obesity <- JSTAR_tn_m_ds_w1 <- JSTAR_tn_ds_w1
save(obesity, file = "../RData/tn_w1/obesity.RData")
rm(obesity)

obesity <- JSTAR_tn_m_ds_w2   <- JSTAR_tn_ds_w2
save(obesity, file = "../RData/tn_w2/obesity.RData")
rm(obesity)


obesity <- JSTAR_cth_m_ds_w1   <- JSTAR_cth_ds_w1
save(obesity, file = "../RData/cth_w1/obesity.RData")
rm(obesity)


l.hds <- list(JSTAR_aksst_ds_w1 = JSTAR_aksst_ds_w1 , JSTAR_aksst_ds_w2 = JSTAR_aksst_ds_w2, JSTAR_aksst_ds_w3=JSTAR_aksst_ds_w3, JSTAR_tn_ds_w1 = JSTAR_tn_ds_w1, JSTAR_tn_ds_w2 = JSTAR_tn_ds_w2, JSTAR_cth_ds_w1 = JSTAR_cth_ds_w1)

m.hds <- list(JSTAR_aksst_m_ds_w1 = JSTAR_aksst_m_ds_w1 , JSTAR_aksst_m_ds_w2 = JSTAR_aksst_m_ds_w2, JSTAR_aksst_m_ds_w3 = JSTAR_aksst_m_ds_w3, JSTAR_tn_m_ds_w1 = JSTAR_tn_m_ds_w1, JSTAR_tn_m_ds_w2 = JSTAR_tn_m_ds_w2, JSTAR_cth_m_ds_w1 = JSTAR_cth_m_ds_w1)



for(i in 1:length(l.hds)) {
  label(l.hds[[i]][[2]]) <- label(m.hds[[i]][[2]]) <- ds_label # Text value assigned at the DS description, at the top.
    
  m.hds[[i]][[2]] <- car::recode(m.hds[[i]][[2]], "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
  m.hds[[i]][[2]] <- labelled(m.hds[[i]][[2]], labels = c("not obese"=0, "obese"=1, "Respond, died after this wave"=2, "Respond(proxy), died"=3, "No respond, alive"=4, "No respond, died this wave"=5, "No respond, died previous waves"=6, "No respond, dropped from sample"=7))

  nl <- names(l.hds[i])
  #opal.assign.data(o, nl , l.hds[[i]])
  #opal.symbol_import(o, nl , project='_Athlos_Harm_Dataset')  

  ml <- names(m.hds[i])
  #opal.assign.data(o, ml , m.hds[[i]]) 
  #opal.symbol_import(o, ml , project='_Athlos_Harm_Dataset')  

}
```



## Descriptive statistics of the harmonised variable accross populations and waves

Percentages of categories in each harmonised variable by cohort and wave:

```{r summ}

# Table HRS
t.hds <- frq(l.hds[[2]][2])[[1]][,c(1,2)] 
for (i in 1:length(l.hds)){
  t.hds[2+i] <- frq(l.hds[[i]][2])[[1]][,4] 
}
t.hds[13,] <- c("n", "sample size", 
             length(l.hds[[1]][[2]]), length(l.hds[[2]][[2]]) , length(l.hds[[3]][[2]]), length(l.hds[[4]][[2]]), length(l.hds[[5]][[2]]), length(l.hds[[6]][[2]]))
names(t.hds)<-c("val", "label", "5cities_W1", "5cities_W2", "5cities_W3", "2cities_W1", "2cities_W2", "3cities_W1")
kable(t.hds)
```



```{r closeRsession, echo=FALSE}
opal.logout(o)
```
