==> JSTAR/6_Physical_measures/7_Grip_Strength.Rmd <==
---
title: "Grip strength"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_laia.r")
```
<!-- ########################################################## --> 



# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `grip`
* Variable label: `Grip strength in kg`
* Variable description: `Grip strength`
* Domain: `Physical measures`
* Value type: `continuous`

* Comments: `According to a systematic review (Roberts et al, 2011), there are some inconsistencies in the use of maximum or mean grip strength as a summary measure which limit comparison of results between epidemiological studies. With multiple attempts, the maximum grip strength will be greater than the mean value. 
Even though for right handed people, there is a 10% stronger grip in the dominant hand, this is not true for left handed or ambidextrous people, whose grip strength are more or less equal in both hands (Roberts et al, 2011). 
In light of the above, the harmonised variable "grip" is proposed to be the highest score (to the nearest kg) of all hand grip measurements in each study, regardless of hand dominance. This is following a protocol suggested by Roberts et al (2011), based on the American Society of Hand Therapists (ASHT) protocol.`

* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

```{r lab, echo=FALSE}
ds_label <- "Grip strength"
```

# Data process

## aksst

### Wave 1 

#### Study-specific variable description


The study-specific variable elected to be harmonised is:

| **Name** | `f_004_3_b`|`f_005_3_b`|
|-|-|-|
| **Label** | `F-004-3:KG OF FIRST MESUREMENT`| `F-005-3:KG OF SECOND MESUREMENT`|
| **Table name** | `jstar_2007_5cities` |`jstar_2007_5cities` |
| **Description** | | | 

```{r assign_aksst1, echo=FALSE}
opal.assign.table.tibble(o, 'JSTAR_gs','JSTAR_datasets_restricted.jstar_2007_5cities',
                         variables=list('f_004_3_b', 'f_005_3_b'), missings = TRUE)
JSTAR_gs <- opal.execute(o,'JSTAR_gs')
```

```{r local_aksst1, echo=FALSE}
vari <- JSTAR_gs$f_004_3_b
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(JSTAR_gs, aes(f_004_3_b)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency")

vari <- JSTAR_gs$f_005_3_b
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(JSTAR_gs, aes(f_005_3_b)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the higest score of all measurements and the missing values have to be recoded as follows: 

* `NA into 999`

**R script:**

```{r harmo_aksst1}
JSTAR_aksst_ds_w1 <- tibble(id=JSTAR_gs$id)
JSTAR_aksst_ds_w1$grip <- apply(JSTAR_gs[,2:3], MARGIN = 1, FUN = function(x) max(x, na.rm = TRUE))
JSTAR_aksst_ds_w1$grip <- car::recode(JSTAR_aksst_ds_w1$grip, "-Inf=999")
JSTAR_aksst_ds_w1$grip <- as.integer(JSTAR_aksst_ds_w1$grip)
JSTAR_aksst_ds_w1$grip <- labelled(JSTAR_aksst_ds_w1$grip, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_aksst1, echo=FALSE}
vari <- JSTAR_aksst_ds_w1$grip
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(JSTAR_aksst_ds_w1, aes(grip)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100) 
```

#### Validation
```{r crosstabulation_aksst1, echo=FALSE}
```

```{r importOPAL_aksst1, echo=FALSE}
#val_labels(JSTAR_aksst_ds_w1) <- NULL
#opal.assign.data(o, 'JSTAR_aksst_ds_w1', JSTAR_aksst_ds_w1)
#opal.symbol_import(o,'JSTAR_aksst_ds_w1', project='_Athlos_Harm_Dataset')
grip <- JSTAR_aksst_ds_w1
save(grip, file = "../RData/aksst_w1/grip.RData")
rm(grip)
```




### Wave 2

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

| **Name** | `f_004_3_b`|`f_005_3_b`|
|-|-|-|
| **Label** | `F-004-3:KG OF FIRST MESUREMENT`| `F-005-3:KG OF SECOND MESUREMENT`|
| **Table name** | `jstar_2009_5cities` |`jstar_2009_5cities` |
| **Description** | | | 

```{r assign_aksst2, echo=FALSE}
opal.assign.table.tibble(o, 'JSTAR_gs','JSTAR_datasets_restricted.jstar_2009_5cities',
                         variables=list('f_004_3_b', 'f_005_3_b'), missings = TRUE)
JSTAR_gs <- opal.execute(o,'JSTAR_gs')
```

```{r local_aksst2, echo=FALSE}
vari <- JSTAR_gs$f_004_3_b
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(JSTAR_gs, aes(f_004_3_b)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency")

vari <- JSTAR_gs$f_005_3_b
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(JSTAR_gs, aes(f_005_3_b)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the higest score of all measurements and the missing values have to be recoded as follows: 

* `NA into 999`


**R script:**

```{r harmo_aksst2}
JSTAR_aksst_ds_w2 <- tibble(id=JSTAR_gs$id)
JSTAR_aksst_ds_w2$grip <- apply(JSTAR_gs[,2:3], MARGIN = 1, FUN = function(x) max(x, na.rm = TRUE))
JSTAR_aksst_ds_w2$grip <- car::recode(JSTAR_aksst_ds_w2$grip, "-Inf=999")
JSTAR_aksst_ds_w2$grip <- as.integer(JSTAR_aksst_ds_w2$grip)
JSTAR_aksst_ds_w2$grip <- labelled(JSTAR_aksst_ds_w2$grip, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_aksst2, echo=FALSE}
vari <- JSTAR_aksst_ds_w2$grip
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(JSTAR_aksst_ds_w2, aes(grip)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100) 
```

#### Validation
```{r crosstabulation_aksst2, echo=FALSE}
```

```{r importOPAL_aksst2, echo=FALSE}
#val_labels(JSTAR_aksst_ds_w2) <- NULL
#opal.assign.data(o, 'JSTAR_aksst_ds_w2', JSTAR_aksst_ds_w2)
#opal.symbol_import(o,'JSTAR_aksst_ds_w2', project='_Athlos_Harm_Dataset')
grip <- JSTAR_aksst_ds_w2
save(grip, file = "../RData/aksst_w2/grip.RData")
rm(grip)
```





### Wave 3 

#### Study-specific variable description


The study-specific variable elected to be harmonised is:

| **Name** | `f_004_3_b`|`f_005_3_b`|
|-|-|-|
| **Label** | `F-004-3:KG OF FIRST MESUREMENT`| `F-005-3:KG OF SECOND MESUREMENT`|
| **Table name** | `jstar_2007_5cities` |`jstar_2007_5cities` |
| **Description** | | | 

```{r assign_aksst3, echo=FALSE}
opal.assign.table.tibble(o, 'JSTAR_gs','JSTAR_datasets_restricted.jstar_2011_5cities',
                         variables=list('f_004_3_b', 'f_005_3_b'), missings = TRUE)
JSTAR_gs <- opal.execute(o,'JSTAR_gs')
```

```{r local_aksst3, echo=FALSE}
vari <- JSTAR_gs$f_004_3_b
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(JSTAR_gs, aes(f_004_3_b)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency")

vari <- JSTAR_gs$f_005_3_b
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(JSTAR_gs, aes(f_005_3_b)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the higest score of all measurements and the missing values have to be recoded as follows: 

* `NA into 999`

**R script:**

```{r harmo_aksst3}
JSTAR_aksst_ds_w3 <- tibble(id=JSTAR_gs$id)
JSTAR_aksst_ds_w3$grip <- apply(JSTAR_gs[,2:3], MARGIN = 1, FUN = function(x) max(x, na.rm = TRUE))
JSTAR_aksst_ds_w3$grip <- car::recode(JSTAR_aksst_ds_w3$grip, "-Inf=999")
JSTAR_aksst_ds_w3$grip <- as.integer(JSTAR_aksst_ds_w3$grip)
JSTAR_aksst_ds_w3$grip <- labelled(JSTAR_aksst_ds_w3$grip, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_aksst3, echo=FALSE}
vari <- JSTAR_aksst_ds_w3$grip
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(JSTAR_aksst_ds_w3, aes(grip)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100) 
```

#### Validation
```{r crosstabulation_aksst3, echo=FALSE}
```

```{r importOPAL_aksst3, echo=FALSE}
#val_labels(JSTAR_aksst_ds_w3) <- NULL
#opal.assign.data(o, 'JSTAR_aksst_ds_w3', JSTAR_aksst_ds_w3)
#opal.symbol_import(o,'JSTAR_aksst_ds_w3', project='_Athlos_Harm_Dataset')
grip <- JSTAR_aksst_ds_w3
save(grip, file = "../RData/aksst_w3/grip.RData")
rm(grip)
```




## tn

### Wave 1 

#### Study-specific variable description


The study-specific variable elected to be harmonised is:

| **Name** | `f_004_3_b`|`f_005_3_b`|
|-|-|-|
| **Label** | `F-004-3:KG OF FIRST MESUREMENT`| `F-005-3:KG OF SECOND MESUREMENT`|
| **Table name** | `jstar_2009_2cities` |`jstar_2009_2cities` |
| **Description** | | | 

```{r assign_tn1, echo=FALSE}
opal.assign.table.tibble(o, 'JSTAR_gs','JSTAR_datasets_restricted.jstar_2009_2cities',
                         variables=list('f_004_3_b', 'f_005_3_b'), missings = TRUE)
JSTAR_gs <- opal.execute(o,'JSTAR_gs')
```

```{r local_tn1, echo=FALSE}
vari <- JSTAR_gs$f_004_3_b
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(JSTAR_gs, aes(f_004_3_b)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency")

vari <- JSTAR_gs$f_005_3_b
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(JSTAR_gs, aes(f_005_3_b)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the higest score of all measurements and the missing values have to be recoded as follows: 

* `NA into 999`

**R script:**


```{r harmo_tn1}
JSTAR_tn_ds_w1 <- tibble(id=JSTAR_gs$id)
JSTAR_tn_ds_w1$grip <- apply(JSTAR_gs[,2:3], MARGIN = 1, FUN = function(x) max(x, na.rm = TRUE))
JSTAR_tn_ds_w1$grip <- car::recode(JSTAR_tn_ds_w1$grip, "-Inf=999")
JSTAR_tn_ds_w1$grip <- as.integer(JSTAR_tn_ds_w1$grip)
JSTAR_tn_ds_w1$grip <- labelled(JSTAR_tn_ds_w1$grip, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_tn1, echo=FALSE}
vari <- JSTAR_tn_ds_w1$grip
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(JSTAR_tn_ds_w1, aes(grip)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100) 
```

#### Validation
```{r crosstabulation_tn1, echo=FALSE}
```

```{r importOPAL_tn1, echo=FALSE}
#val_labels(JSTAR_tn_ds_w1) <- NULL
#opal.assign.data(o, 'JSTAR_tn_ds_w1', JSTAR_tn_ds_w1)
#opal.symbol_import(o,'JSTAR_tn_ds_w1', project='_Athlos_Harm_Dataset')
grip <- JSTAR_tn_ds_w1
save(grip, file = "../RData/tn_w1/grip.RData")
rm(grip)
```



### Wave 2 

#### Study-specific variable description


The study-specific variable elected to be harmonised is:

| **Name** | `f_004_3_b`|`f_005_3_b`|
|-|-|-|
| **Label** | `F-004-3:KG OF FIRST MESUREMENT`| `F-005-3:KG OF SECOND MESUREMENT`|
| **Table name** | `jstar_2007_5cities` |`jstar_2007_5cities` |
| **Description** | | | 

```{r assign_tn2, echo=FALSE}
opal.assign.table.tibble(o, 'JSTAR_gs','JSTAR_datasets_restricted.jstar_2011_2cities',
                         variables=list('f_004_3_b', 'f_005_3_b'), missings = TRUE)
JSTAR_gs <- opal.execute(o,'JSTAR_gs')
```

```{r local_tn2, echo=FALSE}
vari <- JSTAR_gs$f_004_3_b
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(JSTAR_gs, aes(f_004_3_b)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency")

vari <- JSTAR_gs$f_005_3_b
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(JSTAR_gs, aes(f_005_3_b)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the higest score of all measurements and the missing values have to be recoded as follows: 

* `NA into 999`

**R script:**

```{r harmo_tn2}
JSTAR_tn_ds_w2 <- tibble(id=JSTAR_gs$id)
JSTAR_tn_ds_w2$grip <- apply(JSTAR_gs[,2:3], MARGIN = 1, FUN = function(x) max(x, na.rm = TRUE))
JSTAR_tn_ds_w2$grip <- car::recode(JSTAR_tn_ds_w2$grip, "-Inf=999")
JSTAR_tn_ds_w2$grip <- as.integer(JSTAR_tn_ds_w2$grip)
JSTAR_tn_ds_w2$grip <- labelled(JSTAR_tn_ds_w2$grip, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_tn2, echo=FALSE}
vari <- JSTAR_tn_ds_w2$grip
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(JSTAR_tn_ds_w2, aes(grip)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100) 
```

#### Validation
```{r crosstabulation_tn2, echo=FALSE}
```

```{r importOPAL_tn2, echo=FALSE}
#val_labels(JSTAR_tn_ds_w2) <- NULL
#opal.assign.data(o, 'JSTAR_tn_ds_w2', JSTAR_tn_ds_w2)
#opal.symbol_import(o,'JSTAR_tn_ds_w2', project='_Athlos_Harm_Dataset')
grip <- JSTAR_tn_ds_w2
save(grip, file = "../RData/tn_w2/grip.RData")
rm(grip)
```




## cth

### Wave 1 

#### Study-specific variable description


The study-specific variable elected to be harmonised is:

| **Name** | `f_004_3_b`|`f_005_3_b`|
|-|-|-|
| **Label** | `F-004-3:KG OF FIRST MESUREMENT`| `F-005-3:KG OF SECOND MESUREMENT`|
| **Table name** | `jstar_2007_5cities` |`jstar_2007_5cities` |
| **Description** | | | 

```{r assign_cth1, echo=FALSE}
opal.assign.table.tibble(o, 'JSTAR_gs','JSTAR_datasets_restricted.jstar_2011_3cities',
                         variables=list('f_004_3_b', 'f_005_3_b'), missings = TRUE)
JSTAR_gs <- opal.execute(o,'JSTAR_gs')
```

```{r local_cth1, echo=FALSE}
vari <- JSTAR_gs$f_004_3_b
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(JSTAR_gs, aes(f_004_3_b)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency")

vari <- JSTAR_gs$f_005_3_b
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(JSTAR_gs, aes(f_005_3_b)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the higest score of all measurements and the missing values have to be recoded as follows: 

* `NA into 999`

**R script:**

```{r harmo_cth1}
JSTAR_cth_ds_w1 <- tibble(id=JSTAR_gs$id)
JSTAR_cth_ds_w1$grip <- apply(JSTAR_gs[,2:3], MARGIN = 1, FUN = function(x) max(x, na.rm = TRUE))
JSTAR_cth_ds_w1$grip <- car::recode(JSTAR_cth_ds_w1$grip, "-Inf=999")
JSTAR_cth_ds_w1$grip <- as.integer(JSTAR_cth_ds_w1$grip)
JSTAR_cth_ds_w1$grip <- labelled(JSTAR_cth_ds_w1$grip, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_cth1, echo=FALSE}
vari <- JSTAR_cth_ds_w1$grip
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(JSTAR_cth_ds_w1, aes(grip)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100) 
```

#### Validation
```{r crosstabulation_cth1, echo=FALSE}
```

```{r importOPAL_cth1, echo=FALSE}
#val_labels(JSTAR_cth_ds_w1) <- NULL
#opal.assign.data(o, 'JSTAR_cth_ds_w1', JSTAR_cth_ds_w1)
#opal.symbol_import(o,'JSTAR_cth_ds_w1', project='_Athlos_Harm_Dataset')
grip <- JSTAR_cth_ds_w1
save(grip, file = "../RData/cth_w1/grip.RData")
rm(grip)
```


# Quality control
No comments.




```{r closeRsession, echo=FALSE}
opal.logout(o)
```
