==> JSTAR/6_Physical_measures/11_Diastolic.Rmd <==
---
title: "Diastolic"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_iago.r")
```
<!-- ########################################################## --> 



# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `dbp`
* Variable label: `Diastolic blood pressure of participantat at each wave`
* Variable description: `Means of all diastolic blood pressure measures`
* Domain: `Physical measures`
* Value type: `continuous`
* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

# Data process



### JSTAR - aksst - Wave 1 

There is no variable.






### JSTAR - aksst - Wave 2 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `d2_002_3_b`
* Label: `BLOOD PRESSURE, 1ST MEASUREMENT, DIASTOLIC BP (mmHg)`
* Type of variable: `countinous`

* Name: `d2_002_3_g`
* Label: `BLOOD PRESSURE, 2ND MEASUREMENT, DIASTOLIC BP (mmHg)`
* Type of variable: `countinous`

* Name: `d2_002_3_l`
* Label: `BLOOD PRESSURE, 3RD MEASUREMENT, DIASTOLIC BP (mmHg)`
* Type of variable: `countinous`

```{r assign_aksst2}
opal.assign.table.tibble(o, 'JSTAR_aksst_w2_dbp','JSTAR_datasets_restricted.jstar_2009_5cities',
                         variables=list('d2_002_3_b', 'd2_002_3_g', 'd2_002_3_l'), missings = TRUE)
```

```{r local_aksst2}
JSTAR_aksst_w2_dbp <- opal.execute(o,'JSTAR_aksst_w2_dbp')

Continuous_summary(var = JSTAR_aksst_w2_dbp$d2_002_3_b, missing_values = NA)
ggplot(JSTAR_aksst_w2_dbp, aes(d2_002_3_b)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = JSTAR_aksst_w2_dbp$d2_002_3_g, missing_values = NA)
ggplot(JSTAR_aksst_w2_dbp, aes(d2_002_3_g)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = JSTAR_aksst_w2_dbp$d2_002_3_l, missing_values = NA)
ggplot(JSTAR_aksst_w2_dbp, aes(d2_002_3_l)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the mean value between the first, second and third measurement of the study-specific variables, but the missing values have to be recoded as follows:

* `NA into 999`

R code of the ALGORITHM:

```{r harmo_aksst2}
JSTAR_aksst_w2_dbp$mean_dbp <- apply(X = JSTAR_aksst_w2_dbp[,c(2,3,4)], MARGIN = 1, FUN = mean)

#The individuals with NA in some of the measurement but not all:
Nas_ind <- which(is.na(JSTAR_aksst_w2_dbp$d2_002_3_b) | 
                   is.na(JSTAR_aksst_w2_dbp$d2_002_3_g) |
                   is.na(JSTAR_aksst_w2_dbp$d2_002_3_l) ) 
all_Nas_ind <- which(is.na(JSTAR_aksst_w2_dbp$d2_002_3_b) & 
                       is.na(JSTAR_aksst_w2_dbp$d2_002_3_g) & 
                       is.na(JSTAR_aksst_w2_dbp$d2_002_3_l) )
new_Nas_ind <- Nas_ind[-which(Nas_ind%in%all_Nas_ind)]

#individuals with NA in d2_002_3_l:
new_Nas_ind_l <- new_Nas_ind[1:2]
JSTAR_aksst_w2_dbp$mean_dbp[new_Nas_ind_l] <- apply(X = JSTAR_aksst_w2_dbp[new_Nas_ind_l,c(2,3)], MARGIN = 1, FUN = mean)

#individuals with NA in d2_002_3_l and d2_002_3_g:
new_Nas_ind_lg <- new_Nas_ind[3:4]
JSTAR_aksst_w2_dbp$mean_dbp[new_Nas_ind_lg] <- JSTAR_aksst_w2_dbp$d2_002_3_b[new_Nas_ind_lg]


JSTAR_aksst_ds_w2 <- tibble(id=JSTAR_aksst_w2_dbp$id)
JSTAR_aksst_ds_w2$dbp <- car::recode(JSTAR_aksst_w2_dbp$mean_dbp, "NA=999")
JSTAR_aksst_ds_w2$dbp <- labelled(JSTAR_aksst_ds_w2$dbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_aksst2}
Continuous_summary(var = JSTAR_aksst_ds_w2$dbp, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(JSTAR_aksst_ds_w2, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(20,160) 
```

#### Validation
```{r crosstabulation_aksst2}
```



```{r importOPAL_aksst2}
#val_labels(JSTAR_aksst_ds_w2) <- NULL
#opal.assign.data(o, 'JSTAR_aksst_ds_w2', JSTAR_aksst_ds_w2)
#opal.symbol_import(o,'JSTAR_aksst_ds_w2', project='_Athlos_Harm_Dataset')
dbp <- JSTAR_aksst_ds_w2
save(dbp, file = "../RData/aksst_w2/dbp.RData")
rm(dbp)
```





### JSTAR - aksst - Wave 3 

There is no variable.







### JSTAR - tn - Wave 1 

There is no variable.



### JSTAR - tn - Wave 2 

There is no variable.



### JSTAR - cth - Wave 1 

There is no variable. 


# Quality estimation
The calculation of diastolic bp was reached by obtaining the mean value of the three variables available for this study.



```{r closeRsession, echo=FALSE}
opal.logout(o)
```
