==> JSTAR/6_Physical_measures/12_Arterial_hypertension.Rmd <==
---
title: "Arterial hypertension"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---


```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_laia.r")
```

# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `ah`
* Variable label: `Arterial hypertension`
* Variable description: `Arterial hypertension. Means of all diastolic blood pressure measures.`
* Domain: `Physical Measures`
* Value type: `Categorical`
* Category coding:

**Code** | **Category Label**
-------- | ------------------
0        | no hypertension
1        | hypertension

* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done 
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

```{r lab, echo=FALSE}
ds_label <- "Arterial hypertension"
ds_label_all <- "Arterial hypertension. Means of all diastolic blood pressure measures."
```



# Data process

## aksst

### Wave 1 

#### Study-specific variable description


The study-specific variable elected to be harmonised is:

| **Name** | `h_001_2`|
|-|-|-|
| **Label** | `HAVE A HIGH BLOOD PRESSURE` |
| **Table name** | `jstar_2007_5cities` |
| **Categories** | `0 = Have not been diagnosed or advised to seek care` |
| | `1 = Have been diagnosed or advised to seek care` |

```{r assign_aksst1, echo=FALSE}
opal.assign.table.tibble(o, 'JSTAR_ah','JSTAR_datasets_restricted.jstar_2007_5cities',
                         variables=list('h_001_2'), missings = TRUE)
JSTAR_ah <- opal.execute(o,'JSTAR_ah')
```

```{r local_aksst1, echo=FALSE}
vari <- JSTAR_ah$h_001_2
kable(Categorical_summary(vari, missing_values = NA)[2], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(JSTAR_ah, aes(x=factor(h_001_2))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`

R code of the ALGORITHM:

```{r harmo_aksst1}
JSTAR_aksst_ds_w1 <- tibble(id=JSTAR_ah$id)
JSTAR_aksst_ds_w1$ah <- car::recode(JSTAR_ah$h_001_2, "NA=999")
JSTAR_aksst_ds_w1$ah <- labelled(JSTAR_aksst_ds_w1$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_aksst1, echo=FALSE}
vari <- JSTAR_aksst_ds_w1$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(JSTAR_aksst_ds_w1, aes(x=factor(ah))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation_aksst1, echo=FALSE}
```

```{r importOPAL_aksst1, echo=FALSE}
#val_labels(JSTAR_aksst_ds_w1) <- NULL
#opal.assign.data(o, 'JSTAR_aksst_ds_w1', JSTAR_aksst_ds_w1)
#opal.symbol_import(o,'JSTAR_aksst_ds_w1', project='_Athlos_Harm_Dataset')
ah <- JSTAR_aksst_ds_w1
save(ah, file = "../RData/aksst_w1/ah.RData")
rm(ah)
```




### Wave 2

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

| **Name** | `sbp`|`dbp`|
|-|-|-|
| **Label** | `Systolic blood pressure of participant at each wave` |`Diastolic blood pressure of participantat at each wave` |
| **Table name** | `sbp` |`dbp` |
| **Categories** | `continuous` |`continuous` |
| **Missings** |`991 = CAPI/interviewer error` | 
| | `992 = Impute` | `992 = Impute` |
| | `993 = Disable to measure` | `993 = Disable to measure` | 
| | `994 = Not attempt/not done` | `994 = Not attempt/not done` |
| | `995 = Does not answer` | `995 = Does not answer` |
| | `996 = Not applicable`| `996 = Not applicable`| 
| | `997 = Refuse` | `997 = Refuse` |
| | `998 = Do not know` | `998 = Do not know` |
| | `999 = Missing`| `999 = Missing`|
| **Description** |  |  |

```{r assign_aksst2, echo=FALSE}
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/JSTAR/RData/aksst_w2/sbp.RData")
load("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/JSTAR/RData/aksst_w2/dbp.RData")
JSTAR_ah <- merge(sbp, dbp, by = "id")
```

```{r local_aksst2, echo=FALSE}
vari <- JSTAR_ah$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(JSTAR_ah, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Siastolic Blood Pressure") + ylab("Frequency") + xlim(75,230)

vari <- JSTAR_ah$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(JSTAR_ah, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(40,140)
```

#### Harmonisation algorithm
To compute the harmonized variable, first of all, it has to be created two categorical variables as follows:

* dh:
  + `dbp >= 90 into 1`
  + `dbp < 90 into 0`

* sh:
  + `sbp >= 140 into 1`
  + `sbp < 140 into 0`

After that, to compute the harmonized variable from this two variables has to be recoded as follows:

* `if dh = 0 and sh = 0 into 0`
* `if dh = 1 and sh = 0 into 1`
* `if dh = 0 and sh = 1 into 1`
* `if dh = 1 and sh = 1 into 1`
* `NA into 999`

**R script:**

```{r harmo_aksst2}
JSTAR_ah$dh <- car::recode(JSTAR_ah$dbp, "0:89.99999999=0; 90:900=1")
JSTAR_ah$sh <- car::recode(JSTAR_ah$sbp, "0:139.9999999=0; 140:900=1")

JSTAR_aksst_ds_w2 <- tibble(id=JSTAR_ah$id)
JSTAR_aksst_ds_w2$ah <- NA
JSTAR_aksst_ds_w2$ah[which(JSTAR_ah$dh==0 & JSTAR_ah$sh==0)] <- 0
JSTAR_aksst_ds_w2$ah[which(JSTAR_ah$dh==1 & JSTAR_ah$sh==0)] <- 1
JSTAR_aksst_ds_w2$ah[which(JSTAR_ah$dh==0 & JSTAR_ah$sh==1)] <- 1
JSTAR_aksst_ds_w2$ah[which(JSTAR_ah$dh==1 & JSTAR_ah$sh==1)] <- 1
JSTAR_aksst_ds_w2$ah[which(JSTAR_ah$dh==999 & JSTAR_ah$sh==999)] <- 999

JSTAR_aksst_ds_w2$ah <- labelled(JSTAR_aksst_ds_w2$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_aksst2, echo=FALSE}
vari <- JSTAR_aksst_ds_w2$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(JSTAR_aksst_ds_w2, aes(x=factor(ah))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation_aksst2, echo=FALSE}
```

```{r importOPAL_aksst2, echo=FALSE}
#val_labels(JSTAR_aksst_ds_w2) <- NULL
#opal.assign.data(o, 'JSTAR_aksst_ds_w2', JSTAR_aksst_ds_w2)
#opal.symbol_import(o,'JSTAR_aksst_ds_w2', project='_Athlos_Harm_Dataset')
ah <- JSTAR_aksst_ds_w2
save(ah, file = "../RData/aksst_w2/ah.RData")
rm(ah)
```





### Wave 3 

#### Study-specific variable description


The study-specific variable elected to be harmonised is:

| **Name** | `h_001_2`|
|-|-|-|
| **Label** | `HAVE A HIGH BLOOD PRESSURE` |
| **Table name** | `jstar_2011_5cities` |
| **Categories** | `1 = Newly diagnosed with or indicated` |
| | `2 = Fully recovered once but recurred in the past 2 years` |
| | `3 = Still be treated` |
| | `4 = Fully recovered/Never diagnosed with/indicated)` |

```{r assign_aksst3, echo=FALSE}
opal.assign.table.tibble(o, 'JSTAR_ah','JSTAR_datasets_restricted.jstar_2011_5cities',
                         variables=list('h_001_2'), missings = TRUE)
JSTAR_ah <- opal.execute(o,'JSTAR_ah')
```

```{r local_aksst3, echo=FALSE}
vari <- JSTAR_ah$h_001_2
kable(Categorical_summary(vari, missing_values = NA)[2], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(JSTAR_ah, aes(x=factor(h_001_2))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonized variable from the study-specific variable it has to be recoded as follows:

* `keep 1 into 1`
* `2 into 1`
* `3 into 1`
* `4 into 0`
* `NA into 999`

R code of the ALGORITHM:

```{r harmo_aksst3}
JSTAR_aksst_ds_w3 <- tibble(id=JSTAR_ah$id)
JSTAR_aksst_ds_w3$ah <- car::recode(JSTAR_ah$h_001_2, "2=1; 3=1; 4=0; NA=999")
JSTAR_aksst_ds_w3$ah <- labelled(JSTAR_aksst_ds_w3$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_aksst3, echo=FALSE}
vari <- JSTAR_aksst_ds_w3$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(JSTAR_aksst_ds_w3, aes(x=factor(ah))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation_aksst3, echo=FALSE}
```

```{r importOPAL_aksst3, echo=FALSE}
#val_labels(JSTAR_aksst_ds_w3) <- NULL
#opal.assign.data(o, 'JSTAR_aksst_ds_w3', JSTAR_aksst_ds_w3)
#opal.symbol_import(o,'JSTAR_aksst_ds_w3', project='_Athlos_Harm_Dataset')
ah <- JSTAR_aksst_ds_w3
save(ah, file = "../RData/aksst_w3/ah.RData")
rm(ah)
```




## tn

### Wave 1 

#### Study-specific variable description


The study-specific variable elected to be harmonised is:

| **Name** | `h_001_2`|
|-|-|-|
| **Label** | `HAVE A HIGH BLOOD PRESSURE` |
| **Table name** | `jstar_2009_2cities` |
| **Categories** | `0 = Have not been diagnosed or advised to seek care` |
| | `1 = Have been diagnosed or advised to seek care` |

```{r assign_tn1, echo=FALSE}
opal.assign.table.tibble(o, 'JSTAR_ah','JSTAR_datasets_restricted.jstar_2009_2cities',
                         variables=list('h_001_2'), missings = TRUE)
JSTAR_ah <- opal.execute(o,'JSTAR_ah')
```

```{r local_tn1, echo=FALSE}
vari <- JSTAR_ah$h_001_2
kable(Categorical_summary(vari, missing_values = NA)[2], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(JSTAR_ah, aes(x=factor(h_001_2))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`

R code of the ALGORITHM:

```{r harmo_tn1}
JSTAR_tn_ds_w1 <- tibble(id=JSTAR_ah$id)
JSTAR_tn_ds_w1$ah <- car::recode(JSTAR_ah$h_001_2, "NA=999")
JSTAR_tn_ds_w1$ah <- labelled(JSTAR_tn_ds_w1$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_tn1, echo=FALSE}
vari <- JSTAR_tn_ds_w1$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(JSTAR_tn_ds_w1, aes(x=factor(ah))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation_tn1, echo=FALSE}
```

```{r importOPAL_tn1, echo=FALSE}
#val_labels(JSTAR_tn_ds_w1) <- NULL
#opal.assign.data(o, 'JSTAR_tn_ds_w1', JSTAR_tn_ds_w1)
#opal.symbol_import(o,'JSTAR_tn_ds_w1', project='_Athlos_Harm_Dataset')
ah <- JSTAR_tn_ds_w1
save(ah, file = "../RData/tn_w1/ah.RData")
rm(ah)
```



### Wave 2 

#### Study-specific variable description


The study-specific variable elected to be harmonised is:

| **Name** | `h_001_2`|
|-|-|-|
| **Label** | `HAVE A HIGH BLOOD PRESSURE` |
| **Table name** | `jstar_2011_2cities` |
| **Categories** | `1 = Newly diagnosed with or indicated` |
| | `2 = Fully recovered once but recurred in the past 2 years` |
| | `3 = Still be treated` |
| | `4 = Fully recovered/Never diagnosed with/indicated)` |

```{r assign_tn2, echo=FALSE}
opal.assign.table.tibble(o, 'JSTAR_ah','JSTAR_datasets_restricted.jstar_2011_2cities',
                         variables=list('h_001_2'), missings = TRUE)
JSTAR_ah <- opal.execute(o,'JSTAR_ah')
```

```{r local_tn2, echo=FALSE}
vari <- JSTAR_ah$h_001_2
kable(Categorical_summary(vari, missing_values = NA)[2], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(JSTAR_ah, aes(x=factor(h_001_2))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonized variable from the study-specific variable it has to be recoded as follows:

* `keep 1 into 1`
* `2 into 1`
* `3 into 1`
* `4 into 0`
* `NA into 999`

R code of the ALGORITHM:

```{r harmo_tn2}
JSTAR_tn_ds_w2 <- tibble(id=JSTAR_ah$id)
JSTAR_tn_ds_w2$ah <- car::recode(JSTAR_ah$h_001_2, "2=1; 3=1; 4=0; NA=999")
JSTAR_tn_ds_w2$ah <- labelled(JSTAR_tn_ds_w2$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_tn2, echo=FALSE}
vari <- JSTAR_tn_ds_w2$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(JSTAR_tn_ds_w2, aes(x=factor(ah))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation_tn2, echo=FALSE}
```

```{r importOPAL_tn2, echo=FALSE}
#val_labels(JSTAR_tn_ds_w2) <- NULL
#opal.assign.data(o, 'JSTAR_tn_ds_w2', JSTAR_tn_ds_w2)
#opal.symbol_import(o,'JSTAR_tn_ds_w2', project='_Athlos_Harm_Dataset')
ah <- JSTAR_tn_ds_w2
save(ah, file = "../RData/tn_w2/ah.RData")
rm(ah)
```




## cth

### Wave 1 

#### Study-specific variable description


The study-specific variable elected to be harmonised is:

| **Name** | `h_001_2`|
|-|-|-|
| **Label** | `HAVE A HIGH BLOOD PRESSURE` |
| **Table name** | `jstar_2011_3cities` |
| **Categories** | `0 = Have not been diagnosed or advised to seek care` |
| | `1 = Have been diagnosed or advised to seek care` |

```{r assign_cth1, echo=FALSE}
opal.assign.table.tibble(o, 'JSTAR_ah','JSTAR_datasets_restricted.jstar_2011_3cities',
                         variables=list('h_001_2'), missings = TRUE)
JSTAR_ah <- opal.execute(o,'JSTAR_ah')
```

```{r local_cth1, echo=FALSE}
vari <- JSTAR_ah$h_001_2
kable(Categorical_summary(vari, missing_values = NA)[2], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(JSTAR_ah, aes(x=factor(h_001_2))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`

R code of the ALGORITHM:

```{r harmo_cth1}
JSTAR_cth_ds_w1 <- tibble(id=JSTAR_ah$id)
JSTAR_cth_ds_w1$ah <- car::recode(JSTAR_ah$h_001_2, "NA=999")
JSTAR_cth_ds_w1$ah <- labelled(JSTAR_cth_ds_w1$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_cth1, echo=FALSE}
vari <- JSTAR_cth_ds_w1$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(JSTAR_cth_ds_w1, aes(x=factor(ah))) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") 
```

#### Validation
```{r crosstabulation_cth1, echo=FALSE}
```

```{r importOPAL_cth1, echo=FALSE}
#val_labels(JSTAR_cth_ds_w1) <- NULL
#opal.assign.data(o, 'JSTAR_cth_ds_w1', JSTAR_cth_ds_w1)
#opal.symbol_import(o,'JSTAR_cth_ds_w1', project='_Athlos_Harm_Dataset')
ah <- JSTAR_cth_ds_w1
save(ah, file = "../RData/cth_w1/ah.RData")
rm(ah)
```

# Quality estimation


No comments.


```{r closeRsession, echo=FALSE}
opal.logout(o)
```
